requirements.txt
flask~=2.2.2
pymongo[srv]~=4.2.0
gunicorn
dnspython~=2.2.1